moved ModuleInAppHookProvider to a separate class
