+++ messaginginapp/src/main/java/io/customer/messaginginapp/ModuleMessagingInApp.kt
@@ -2,14 +2,13 @@ package io.customer.messaginginapp
 
 import android.app.Application
 import io.customer.messaginginapp.di.gistProvider
+import io.customer.messaginginapp.hook.ModuleInAppHookProvider
 import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOModule
 import io.customer.sdk.data.request.MetricEvent
 import io.customer.sdk.di.CustomerIOComponent
-import io.customer.sdk.hooks.HookModules
+import io.customer.sdk.hooks.HookModule
 import io.customer.sdk.hooks.HooksManager
-import io.customer.sdk.hooks.ModuleHook
-import io.customer.sdk.hooks.ModuleHookProvider
 import io.customer.sdk.repository.TrackRepository
 
 class ModuleMessagingInApp internal constructor(
@@ -52,20 +51,8 @@ class ModuleMessagingInApp internal constructor(
 
     private fun setupHooks() {
         hooksManager.add(
-            HookModules.MessagingInApp,
-            object : ModuleHookProvider() {
-                override fun profileIdentifiedHook(hook: ModuleHook.ProfileIdentifiedHook) {
-                    gistProvider.setUserToken(hook.identifier)
-                }
-
-                override fun screenTrackedHook(hook: ModuleHook.ScreenTrackedHook) {
-                    gistProvider.setCurrentRoute(hook.screen)
-                }
-
-                override fun beforeProfileStoppedBeingIdentified(hook: ModuleHook.BeforeProfileStoppedBeingIdentified) {
-                    gistProvider.clearUserToken()
-                }
-            }
+            module = HookModule.MessagingInApp,
+            subscriber = ModuleInAppHookProvider()
         )
     }
